% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADL.R
\name{ADL}
\alias{ADL}
\title{ADL}
\usage{
ADL(close, low, high, volume)
}
\arguments{
\item{close}{vector of closing prices}

\item{low}{vector of low prices}

\item{high}{vector of high prices}

\item{volume}{vector of volumes}
}
\value{
A vector containing the ADL value
}
\description{
Compute the Accumulation/Distribution Line Indicator
}
\examples{

data(TWTR)
ADL(TWTR$Close, TWTR$Low, TWTR$High, TWTR$Volume)

data(BAC)
ADL(BAC$Close, BAC$Low, BAC$High, BAC$Volume)

}
\references{
Tom Taulli, 
\emph{The streetsmart guide to short selling: techniques the pros use 
to profit in any market. McGraw-Hill Professional. p. 190.} 2002.
}
\author{
Paolo Montemurro <montep@usi.ch>
}
