% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aroon.R
\name{Aroon}
\alias{Aroon}
\title{Aroon}
\usage{
Aroon(price, n = 14)
}
\arguments{
\item{price}{vector of prices}

\item{n}{integer, length of each period of analysis}
}
\value{
A dataframe containing aroonUp and aroonDown values
}
\description{
Compute the Aroon indicator
}
\examples{

data(TWTR)
Aroon(TWTR$Close, 5)

}
\references{
Tushar Chande, 
\emph{The New Technical Trader: Boost Your Profit by Plugging 
Into the Latest Indicators}, 1994.
}
\author{
Paolo Montemurro <montep@usi.ch>
}
