% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AroonOscillator.R
\name{AroonOscillator}
\alias{AroonOscillator}
\title{Aroon Oscillator}
\usage{
AroonOscillator(price, n = 14)
}
\arguments{
\item{price}{vector of prices}

\item{n}{integer, length of each period of analysis}
}
\value{
A vector containing aroonOscillator values
}
\description{
Compute the Aroon Oscillator Indicator
}
\examples{
p <- c( 20, 22, 24, 25, 23, 26, 28, 26, 29, 27, 28, 30, 27, 29, 28 )
AroonOscillator(p,5)
}
\references{
Tushar Chande, 
\emph{The New Technical Trader: Boost Your Profit by Plugging Into 
the Latest Indicators}, 1994.
}
\author{
Paolo Montemurro <montep@usi.ch>
}
