% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATR.R
\name{ATR}
\alias{ATR}
\title{Average True Range}
\usage{
ATR(high, low, close, n = 14)
}
\arguments{
\item{high}{vector of high prices}

\item{low}{vector of low prices}

\item{close}{vector of closing prices}

\item{n}{integer, length of each period of analysis}
}
\value{
a vector containing the Average True Range values
}
\description{
Compute the Average True Range Indicator
}
\examples{

data(BAC)
ATR(BAC$High,BAC$Low,BAC$Close)

}
\references{
J. Welles Wilder, 
\emph{New Concepts in Technical Analysis Systems}, 1974.
}
\author{
Paolo Montemurro <montep@usi.ch>
}
