% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAC-data.R
\docType{data}
\name{BAC}
\alias{BAC}
\title{Bank of America Stock Price Data}
\format{An object of class \code{"xts"}.}
\source{
\href{https://finance.yahoo.com/quote/BAC?p=BAC}{Yahoo Finance}
}
\usage{
data(BAC)
}
\description{
Bank of America Corporation (BAC) historical price time series from 
2013-01-02 to 2018-12-31. Data saved in xts class as downloaded with 
quantmod package. Use to test function's compatibility with quantmod
output.
}
\examples{
data(BAC)
head(BAC)
}
\keyword{America}
\keyword{BAC,}
\keyword{Bank}
\keyword{datasets,}
\keyword{of}
