% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bollinger.R
\name{Bollinger}
\alias{Bollinger}
\title{Bollinger Bands}
\usage{
Bollinger(price, period = 20, n = 2)
}
\arguments{
\item{price}{stock price time serie, either a vector or an xts object}

\item{period}{smoothing period of the moving average, an integer or a 
vector of integers}

\item{n}{number of standard deviations, an integer or a vector of integers}
}
\value{
a dataframe or an \code{\link[xts:xts-package]{xts}} object 
depending on the input, composed of 3 columns: the simple moving average
(\code{SMA}), the Upper Bollinger Band value (\code{UpperBB}) and the 
Lower Bollinger Band value (\code{LowerBB}). If \code{period} or \code{n}
are multivalued then a list contained the output described above for each
value.
}
\description{
Calculates a stock Bollinger Bands based on its volatility.
}
\examples{

Bollinger( USGDP )

}
\references{
Murphy John J. (1999),
\emph{Technical Analysis of the Financial Markets},
New York Institute of Finance.
}
\seealso{
Moving Average function: \code{\link{MA}}.
}
\author{
Giovanni Kraushaar <giovanni.kraushaar@usi.ch>
}
