% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOP.R
\name{BOP}
\alias{BOP}
\title{BOP}
\usage{
BOP(close, open, high, low)
}
\arguments{
\item{close}{vector of closing prices}

\item{open}{vector of opening prices}

\item{high}{vector of high prices}

\item{low}{vector of low prices}
}
\value{
A vector containing the BOP value
}
\description{
Compute the Balance of Power Indicator
}
\examples{

data(TWTR)
BOP(TWTR$Close, TWTR$Open, TWTR$High, TWTR$Low)

data(BAC)
BOP(BAC$Close, BAC$Open, BAC$High, BAC$Low)

}
\references{
Igor Livshin, \emph{Balance of market power}.
}
\author{
Paolo Montemurro <montep@usi.ch>
}
