% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMV.R
\name{EMV}
\alias{EMV}
\title{EMV}
\usage{
EMV(high, low, volume, divisor = 10000)
}
\arguments{
\item{high}{vector of high prices}

\item{low}{vector of low prices}

\item{volume}{vector of volumes}

\item{divisor}{integer that scales the result}
}
\value{
A vector containing the EMV value
}
\description{
Compute the Arms Ease of Movement
}
\examples{

data(TWTR)
EMV(TWTR$High, TWTR$Low, TWTR$Volume)

data(BAC)
EMV(BAC$High, BAC$Low, BAC$Volume)

}
\references{
Richard W. Arms, Jr, \emph{Arms Ease of Movement},
}
\author{
Paolo Montemurro <montep@usi.ch>
}
