% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ichimoku.R
\name{Ichimoku}
\alias{Ichimoku}
\title{Ichimoku Kinko Hyo}
\usage{
Ichimoku(high, low, closingPrice)
}
\arguments{
\item{high}{A vector of high prices or an xts object.}

\item{low}{A vector of low prices or an xts object.}

\item{closingPrice}{A vector of past closing prices or an xts object.}
}
\value{
A matrix or an xts object, accordingly to the input, of the same 
length of the input.
}
\description{
Compute the Ichimoku Kinko Hyo indicator.
}
\examples{

data(BAC)
Ichimoku( BAC$High, BAC$Low, BAC$Close)

}
\references{
Goichi Hosoda, 1930-69
}
\author{
Luca Sanfilippo <luca.sanfilippo@usi.ch>
}
