% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MA.R
\name{MA}
\alias{MA}
\title{Moving Average}
\usage{
MA(price, period, method = "SMA")
}
\arguments{
\item{price}{A vector of past prices.}

\item{period}{Integer or vector of integers. Number of datapoints over 
which calculate the average.}

\item{method}{One of \code{c('SMA', 'EMA', 'WMA')}, for simple-, 
exponential-, or weighted- moving average.}
}
\description{
Moving Average
}
\examples{
# prices vector
p <- c( 20, 22, 24, 25, 23, 26, 28, 26, 29, 27, 28, 30, 27, 29, 28 ) 
MA(p, method='sma', period = c(10,2))
}
\author{
Giovanni Kraushaar <giovanni.kraushaar@usi.ch>
}
