% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACD.R
\name{MACD}
\alias{MACD}
\title{Moving Average Convergence Divergence}
\usage{
MACD(price, period = c(12, 26), signal_line = FALSE, slp = 9,
  k = NULL)
}
\arguments{
\item{price}{A vector of past prices or an xts object.}

\item{period}{\code{c(sp,lp)}, a vector containing the length of the short
period EMA and the length of the long period EMA.}

\item{signal_line}{Boolean, if true the signal line using a period of 
\code{slp} is computed.}

\item{slp}{Signal line period, the length of the period of the EMA used as
signal line.}

\item{k}{Weighting multiplier for the underlying Exponential
\code{\link{MA}}, defualt if \code{NULL}.}
}
\value{
A \code{\link[base:vector]{vector}} or an 
\code{\link[xts:xts-package]{xts}} object, accordingly to the input, 
of the same length of the input. If also the 
signal line is computed, a list with two elements: \code{MACD} and 
\code{signal}, or a multiple timeseries if it is an xts.
}
\description{
Computes the Moving Average Convergence Divergence (MACD), 
a trend-following momentum indicator that shows the relationship between
two moving averages of a security’s price.
}
\examples{
# Compute Exponential Moving Average of Bank of America quotes
MACD( BAC$Close )

}
\references{
Murphy John J. (1999),
\emph{Technical Analysis of the Financial Markets},
New York Institute of Finance.
}
\seealso{
Moving average: \code{\link{MA}}.
}
\author{
Giovanni Kraushaar <giovanni.kraushaar@usi.ch>
}
