% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OBV.R
\name{OBV}
\alias{OBV}
\title{On Balance Volume}
\usage{
OBV(closingPrice, volume)
}
\arguments{
\item{closingPrice}{A vector or an xts vector of past closing prices.}

\item{volume}{A vector or an xts vector of past volumes}
}
\value{
A vector or an xts object, accordingly to the input, of the same 
length of the input.
}
\description{
Compute the On Balance Volume indicator.
}
\examples{

obv <- OBV(BAC$Close, BAC$Volume)
plot( obv, type = 'l')

}
\references{
Granville Joe (1963), \emph{New Key To Stock Market Profits}.
}
\author{
Luca Sanfilippo <luca.sanfilippo@usi.ch>
}
