% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSI.R
\name{RSI}
\alias{RSI}
\title{RSI}
\usage{
RSI(returns, n = 14)
}
\arguments{
\item{returns}{vector of closing prices}

\item{n}{integer, length of each period of analysis}
}
\value{
a vector containing RSI value
}
\description{
Compute the Relative Strength Index
}
\examples{

data(BAC)
returns <- (BAC$Close / BAC$Open) - 1
RSI(returns)

}
\references{
Welles Wilder, 
\emph{New Concepts in Technical Trading Systems}, 1978
}
\author{
Paolo Montemurro <montep@usi.ch>
}
