% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StochOscillator.R
\name{StochOscillator}
\alias{StochOscillator}
\title{StochOscillator}
\usage{
StochOscillator(close, high, low, n = 14, k = 3)
}
\arguments{
\item{close}{vector of closing prices}

\item{high}{vector of high prices}

\item{low}{vector of low prices}

\item{n}{integer, length of each period of analysis}

\item{k}{integer, length of the smoothing factor}
}
\description{
Compute stochastic oscillator
}
\examples{

data(TWTR)
StochOscillator(TWTR$Close,TWTR$High,TWTR$Low)
data(BAC)
StochOscillator(BAC$Close,BAC$High,BAC$Low)

}
\references{
George C. Lane, 
\emph{John J. Murphy, Technical Analysis of the Financial Markets}, 1950.
}
\author{
Paolo Montemurro <montep@usi.ch>
}
