% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StochRSI.R
\name{StochRSI}
\alias{StochRSI}
\title{Stochastic Relative Strenght Index}
\usage{
StochRSI(rsi, period = 14)
}
\arguments{
\item{rsi}{A vector of the relative strength index values calculated 
with a period of 14 days or an xts object.}

\item{period}{It is a value (in days: usually it is 14days)}
}
\value{
A vector or an xts object, accordingly to the input, of the same 
length of the input.
}
\description{
Compute the Stochastic Relative Strenght Index indicator.
}
\examples{

data(BAC)
rsi <- RSI(BAC$Close)
StochRSI(rsi)
}
\references{
Chande Tushard  and Kroll Stanley (1994), 
\emph{The New Technical Trader}.
}
\author{
Luca Sanfilippo <luca.sanfilippo@usi.ch>
}
