% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TWTR-data.R
\docType{data}
\name{TWTR}
\alias{TWTR}
\title{Twitter Stock Prices Data}
\format{An object of class \code{"data.frame"}.}
\source{
\href{https://finance.yahoo.com/quote/TWTR/history?period1=1524528000&period2=1556064000&interval=1d&filter=history&frequency=1d}{Yahoo Finance}
}
\usage{
data(TWTR)
}
\description{
Historical data on daily Twitter, Inc. (TWTR) stock prices between 
2018-04-24 and 2019-04-23 in USD. Retrieved from Yahoo Finance on 
April 24th, 2019.
}
\details{
Columns: \code{Date}, \code{Open}, \code{High}, \code{Low}, 
\code{Close}, \code{Adj.Close}, \code{Volume}
}
\examples{
data(TWTR)
str(TWTR)
head(TWTR)
}
\keyword{TWTR}
\keyword{datasets,}
\keyword{twitter,}
