% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USGDP-data.R
\docType{data}
\name{USGDP}
\alias{USGDP}
\title{US Real GDP Data}
\format{An object of class \code{"xts"}.}
\source{
\href{https://fred.stlouisfed.org/series/GDPC1}{FRED}
}
\usage{
data(USGDP)
}
\description{
United States' Real Gross Domestic Product in billions of chained 2012
US dollars, seasonally adjusted annual rate, from 1947-01-01 to 2018-10-01
with quarterly frequency.
}
\examples{
data(USGDP)
head(USGDP)
}
\references{
U.S. Bureau of Economic Analysis, \emph{Real Gross Domestic 
Product} "GDPC1", retrieved from FRED, Federal Reserve Bank of St. Louis; 
\code{https://fred.stlouisfed.org/series/GDPC1}, April 24, 2019.
}
\keyword{GDP,}
\keyword{USGDP}
\keyword{datasets,}
