% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADL.R
\name{ADL}
\alias{ADL}
\title{ADL}
\usage{
ADL(close, low, high, volume)
}
\arguments{
\item{close}{vector or xts, historical series of the close price for each interval}

\item{low}{vector or xts, historical series of the lowest prices for each interval}

\item{high}{vector or xts, historical series of the highest prices for each interval}

\item{volume}{vector or xts, historical series of volume for each interval}
}
\value{
vector or xts, historical series of ADL value
}
\description{
Compute the Accumulation/Distribution Line Indicator
}
\examples{

data(TWTR)
ADL(TWTR$Close, TWTR$Low, TWTR$High, TWTR$Volume)

data(BAC)
ADL(BAC$Close, BAC$Low, BAC$High, BAC$Volume)

}
\references{
Taulli Tom (2002), 
\emph{The streetsmart guide to short selling: 
techniques the pros use to profit in any market}. 
McGraw-Hill Professional. p. 190.
}
\author{
Paolo Montemurro <montep@usi.ch>
}
