% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aroon.R
\name{Aroon}
\alias{Aroon}
\title{Aroon}
\usage{
Aroon(price, n = 14)
}
\arguments{
\item{price}{vector or xts, historical series of prices}

\item{n}{integer, length of each period of analysis}
}
\value{
dataframe or xts, historical series of AroonUp and AroonDown values
}
\description{
Compute the Aroon indicator
}
\examples{

data(TWTR)
Aroon(TWTR$Close, 5)

}
\references{
Chande Tushar (1994), 
\emph{The New Technical Trader: 
Boost Your Profit by Plugging Into the Latest Indicators}.
}
\author{
Paolo Montemurro <montep@usi.ch>
}
