% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AroonOscillator.R
\name{AroonOscillator}
\alias{AroonOscillator}
\title{Aroon Oscillator}
\usage{
AroonOscillator(price, n = 14)
}
\arguments{
\item{price}{vector or xts, historical series of prices}

\item{n}{integer, length of each period of analysis}
}
\value{
vector or xts, historical series of Aroon Oscillator value
}
\description{
Compute the Aroon Oscillator Indicator
}
\examples{

p <- c( 20, 22, 24, 25, 23, 26, 28, 26, 29, 27, 28, 30, 27, 29, 28 )
AroonOscillator(p,5)

}
\references{
Chande Tushar (1994), \emph{The New Technical Trader: 
Boost Your Profit by Plugging Into the Latest Indicators}.
}
\author{
Paolo Montemurro <montep@usi.ch>
}
