% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATR.R
\name{ATR}
\alias{ATR}
\title{Average True Range}
\usage{
ATR(close, low, high, n = 14)
}
\arguments{
\item{close}{vector or xts, historical series of the close price for each interval}

\item{low}{vector or xts, historical series of the lowest prices for each interval}

\item{high}{vector or xts, historical series of the highest prices for each interval}

\item{n}{integer, length of each period of analysis}
}
\value{
vector or xts, historical series of ATR value
}
\description{
Compute the Average True Range Indicator
}
\examples{

data(BAC)
ATR(BAC$Close,BAC$Low,BAC$High)

}
\references{
Wilder J. Welles (1974), 
\emph{New Concepts in Technical Analysis Systems}.
}
\author{
Paolo Montemurro <montep@usi.ch>
}
