% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOP.R
\name{BOP}
\alias{BOP}
\title{BOP}
\usage{
BOP(open, close, high, low)
}
\arguments{
\item{open}{vector or xts, historical series of the open price for each interval}

\item{close}{vector or xts, historical series of the close price for each interval}

\item{high}{vector or xts, historical series of the highest prices for each interval}

\item{low}{vector or xts, historical series of the lowest prices for each interval}
}
\value{
vector or xts, historical series of BOP value
}
\description{
Compute the Balance of Power Indicator
}
\examples{

data(TWTR)
BOP(TWTR$Open, TWTR$Close, TWTR$High, TWTR$Low)

data(BAC)
BOP(BAC$Open, BAC$Close, BAC$High, BAC$Low)

}
\references{
Igor Livshin, \emph{Balance of market power}.
}
\author{
Paolo Montemurro <montep@usi.ch>
}
