% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMV.R
\name{EMV}
\alias{EMV}
\title{EMV}
\usage{
EMV(low, high, volume, divisor = 10000)
}
\arguments{
\item{low}{vector or xts, historical series of the lowest prices for each interval}

\item{high}{vector or xts, historical series of the highest prices for each interval}

\item{volume}{vector or xts, historical series of volume for each interval}

\item{divisor}{integer that scales the result}
}
\value{
vector or xts, historical series of EMV value
}
\description{
Compute the Arms Ease of Movement
}
\examples{

data(TWTR)
EMV(TWTR$Low, TWTR$High, TWTR$Volume)

data(BAC)
EMV(BAC$Low, BAC$High, BAC$Volume)

}
\references{
Richard W. Arms, Jr, \emph{Arms Ease of Movement},
}
\author{
Paolo Montemurro <montep@usi.ch>
}
