% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormPrice.R
\name{NormPrice}
\alias{NormPrice}
\title{Normalized Price}
\usage{
NormPrice(price, scale = 100)
}
\arguments{
\item{price}{vector or xts, historical series of prices}

\item{scale}{integer, scale of the result}
}
\value{
vector or xts, historical series of normalized prices to the inserted scale
}
\description{
Compute the normalized price
}
\examples{

p <- c( 20, 22, 24, 25, 23, 26, 28, 26, 29, 27, 28, 30, 27, 29, 28 )
NormPrice(p,1)

}
\author{
Paolo Montemurro <montep@usi.ch>
}
