% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StochOscillator.R
\name{StochOscillator}
\alias{StochOscillator}
\title{StochOscillator}
\usage{
StochOscillator(close, high, low, n = 14, k = 3)
}
\arguments{
\item{close}{vector or xts, historical series of the close price for each interval}

\item{high}{vector or xts, historical series of the highest prices for each interval}

\item{low}{vector or xts, historical series of the lowest prices for each interval}

\item{n}{integer, length of each period of analysis}

\item{k}{integer, length of the smoothing factor}
}
\value{
vector or xts, historical series of stochastic oscillator value
}
\description{
Compute stochastic oscillator
}
\examples{

data(TWTR)
StochOscillator(TWTR$Close,TWTR$Low,TWTR$High)
data(BAC)
StochOscillator(BAC$Close,BAC$Low,BAC$High)

}
\references{
George C. Lane, 
\emph{}, 1950.
}
\author{
Paolo Montemurro <montep@usi.ch>
}
