\name{TradingIndicatoR}
\alias{TIR}
\alias{TradingIndicatoR}
\docType{package}
\title{
\packageTitle{TradingIndicatoR}
}
\description{
\packageDescription{TradingIndicatoR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{TradingIndicatoR}
\packageIndices{TradingIndicatoR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{TradingIndicatoR}

Maintainer: \packageMaintainer{TradingIndicatoR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ trading, finance, tchnical trading, trading analysis }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
